Each Class written in separate file and new functions are also added in these classes
